refactor to not mess up process...
